[ENH] Subsurface compatibility for gempy output
