+++ gempy/core/solution.py
@@ -1,13 +1,18 @@
 import numpy as np
-from typing import Union, Iterable
+from typing import Union
 import warnings
 from skimage import measure
 from gempy.utils.input_manipulation import find_interfaces_from_block_bottoms
 from gempy.core.data import Grid, Surfaces
 from gempy.core.data_modules.stack import Series, Stack
-from gempy.utils.meta import _setdoc, _setdoc_pro
+from gempy.utils.meta import _setdoc_pro
 import gempy.utils.docstring as ds
-import xarray as xr
+
+try:
+    from gempy.core.xsolution import XSolution
+    inheritance = XSolution
+except ImportError:
+    inheritance = object
 
 
 @_setdoc_pro(
@@ -15,304 +20,7 @@ import xarray as xr
      ds.bai, ds.mai, ds.vai,
      ds.lith_block, ds.sfm, ds.bm, ds.mm, ds.vm, ds.vertices, ds.edges,
      ds.geological_map])
-class XSolution(object):
-    """This class stores the output of the interpolation and the necessary objects
-    to visualize and manipulate this data.
-
-    Depending on the activated grid (see :class:`Grid`) a different number of
-     properties are returned returned:
-
-    Args:
-        grid (Grid): [s0]
-        surfaces (Surfaces): [s1]
-        series (Series): [s2]
-
-    Attributes:
-        grid (Grid)
-        surfaces (Surfaces)
-        series (Series)
-        weights_vector (numpy.array): [s3]
-        scalar_field_at_surface_points (numpy.array): [s4]
-        block_at_surface_points (numpy.array): [s5]
-        mask_at_surface_points (numpy.array): [s6]
-        values_at_surface_points (numpy.array): [s7]
-        lith_block (numpy.array): [s8]
-        scalar_field_matrix (numpy.array): [s9]
-        block_matrix (numpy.array): [s10]
-        mask_matrix (numpy.array): [s11]
-        mask_matrix_pad (numpy.array): mask matrix padded 2 block in order to guarantee that the layers intersect each
-         other after marching cubes
-        values_matrix (numpy.array): [s12]
-        vertices (list[numpy.array]): [s13]
-        edges (list[numpy.array]): [s14]
-        geological_map (numpy.array): [s15]
-
-    """
-
-    def __init__(self, grid: Grid,
-                 surfaces: Surfaces = None,
-                 stack: Stack = None,
-                 ):
-        # self.additional_data = additional_data
-
-        self.grid = grid
-        #  self.surface_points = surface_points
-        self.stack = stack
-        self.surfaces = surfaces  # Used to store ver/sim there
-
-        # Define xarrays
-        self.weights_vector = None
-        self.at_surface_points = None
-        self.s_regular_grid = xr.Dataset()
-        self.s_custom_grid = xr.Dataset()
-        self.s_topography = xr.Dataset()
-        self.s_at_surface_points = xr.Dataset()
-        self.s_sections = dict()
-        self.meshes = None
-
-    def set_values(self,
-                   values: list,
-                   active_features=None,
-                   surf_properties=None,
-                   attach_xyz=True):
-        """ At this stage we should split values into the different grids
-
-        Args:
-            values:
-
-        Returns:
-
-        """
-
-        # Get an array with all the indices for each grid
-        l = self.grid.length
-
-        coords_base, xyz = self.prepare_common_args(active_features, attach_xyz,
-                                                    surf_properties)
-        self.weights_vector = values[3]
-
-        if self.grid.active_grids[0]:
-            self.set_values_to_regular_grid(values, l[0], l[1], coords_base.copy())
-        if self.grid.active_grids[1]:
-            self.set_values_to_custom_grid(values, l[1], l[2], coords_base.copy(),
-                                           xyz=xyz)
-        if self.grid.active_grids[2]:
-            self.set_values_to_topography(values, l[2], l[3], coords_base.copy())
-        if self.grid.active_grids[3]:
-            self.set_values_to_sections(values, l[3], l[4], coords_base.copy())
-        if self.grid.active_grids[4]:
-            self.set_values_to_centered()
-
-        # TODO: Add xyz from surface points
-        self.set_values_to_surface_points(values, l[-1], coords_base, xyz=None)
-
-    def prepare_common_args(self, active_features, attach_xyz, surf_properties):
-        if active_features is None and self.stack is not None:
-            active_features = self.stack.df.groupby('isActive').get_group(True).index
-        if surf_properties is None and self.surfaces is not None:
-            surf_properties = self.surfaces.properties_val
-        coords_base = dict()
-        if active_features is not None:
-            coords_base['Features'] = active_features
-        if surf_properties is not None:
-            coords_base['Properties'] = surf_properties
-        if attach_xyz and self.grid.custom_grid is not None:
-            xyz = self.grid.custom_grid.values
-        else:
-            xyz = None
-        return coords_base, xyz
-
-    def set_values_to_centered(self):
-        return
-
-    def set_values_to_surface_points(self, values, l0, coords_base, xyz=None):
-        coords = coords_base
-        l1 = values[0].shape[-1]
-        arrays = self.create_unstruct_xarray(values, l0, l1, xyz)
-
-        self.s_at_surface_points = xr.Dataset(
-            data_vars=arrays,
-            coords=coords
-        )
-        return self.s_at_surface_points
-
-    @staticmethod
-    def create_struc_xarrays(values, l0, l1, res: Union[list, np.ndarray]):
-        arrays = dict()
-
-        n_dim = len(res)
-        xyz = ['X', 'Y', 'Z'][:n_dim]
-        if values[0] is not None:
-            # This encompass lith_block and values matrix
-            property_matrix = xr.DataArray(
-                data=values[0][:, l0:l1].reshape(-1, *res),
-                dims=['Properties', *xyz],
-            )
-            arrays['property_matrix'] = property_matrix
-
-        if values[1] is not None:
-            # This is the block matrix
-            i, j, _ = values[1].shape
-            block_matrix = xr.DataArray(
-                data=values[1][:, :, l0:l1].reshape(i, j, *res),
-                dims=['Features', 'Properties', *xyz],
-            )
-            arrays['block_matrix'] = block_matrix
-
-            # Fault block?
-
-        if values[4] is not None:
-            # Scalar field matrix
-            scalar_matrix = xr.DataArray(
-                data=values[4][:, l0:l1].reshape(-1, *res),
-                dims=['Features', *xyz],
-            )
-            arrays['scalar_field_matrix'] = scalar_matrix
-
-        if values[6] is not None:
-            # Mask matrix
-            mask_matrix = xr.DataArray(
-                data=values[6][:, l0:l1].reshape(-1, *res),
-                dims=['Features', *xyz],
-            )
-            arrays['mask_matrix'] = mask_matrix
-
-        if values[7] is not None:
-            # Fault mask matrix
-            fault_mask = xr.DataArray(
-                data=values[7][:, l0:l1].reshape(-1, *res),
-                dims=['Features', *xyz],
-            )
-            arrays['fault_mask'] = fault_mask
-
-        return arrays
-
-    @staticmethod
-    def create_unstruct_xarray(values, l0, l1, xyz):
-        arrays = dict()
-        if xyz is not None:
-            cartesian_matrix = xr.DataArray(
-                data=xyz,
-                dims=['Point', 'XYZ'],
-                coords={'XYZ': ['X', 'Y', 'Z']}
-            )
-            arrays['cartesian_matrix'] = cartesian_matrix
-
-        if values[0] is not None:
-            # Values and lith block
-            property_v3 = xr.DataArray(
-                data=values[0][:, l0:l1],
-                dims=['Properties', 'Point'],
-            )
-
-            arrays['property_v3'] = property_v3
-
-        if values[1] is not None:
-            # block
-            block_v3 = xr.DataArray(
-                data=values[1][:, :, l0:l1],
-                dims=['Features', 'Properties', 'Point'],
-            )
-
-            arrays['block_v3'] = block_v3
-
-        if values[4] is not None:
-            # Scalar field
-            scalar_field_v3 = xr.DataArray(
-                data=values[4][:, l0:l1],
-                dims=['Features', 'Point'],
-            )
-            arrays['scalar_field_v3'] = scalar_field_v3
-
-        if values[6] is not None:
-            # Scalar field
-            mask_v3 = xr.DataArray(
-                data=values[6][:, l0:l1],
-                dims=['Features', 'Point'],
-            )
-            arrays['mask_v3'] = mask_v3
-
-        return arrays
-
-    def set_values_to_custom_grid(self, values: list, l0, l1,
-                                  coords_base: dict, xyz=None):
-
-        coords = coords_base
-        arrays = self.create_unstruct_xarray(values, l0, l1, xyz)
-
-        self.s_custom_grid = xr.Dataset(
-            data_vars=arrays,
-            coords=coords
-        )
-        return self.s_custom_grid
-
-    def set_values_to_regular_grid(self, values: list, l0, l1,
-                                   coords_base: dict):
-
-        coords = self.add_cartesian_coords(coords_base)
-
-        arrays = self.create_struc_xarrays(values, l0, l1,
-                                           self.grid.regular_grid.resolution)
-
-        self.s_regular_grid = xr.Dataset(
-            data_vars=arrays,
-            coords=coords
-        )
-
-    def add_cartesian_coords(self, coords_base):
-        coords = coords_base
-        coords['X'] = self.grid.regular_grid.x
-        coords['Y'] = self.grid.regular_grid.y
-        coords['Z'] = self.grid.regular_grid.z
-        return coords
-
-    def set_values_to_topography(self,
-                                 values: list,
-                                 l0, l1,
-                                 coords_base):
-        coords = coords_base
-        coords['X'] = self.grid.topography.x
-        coords['Y'] = self.grid.topography.y
-        resolution = self.grid.topography.resolution
-        arrays = self.create_struc_xarrays(values, l0, l1, resolution)
-
-        self.s_topography = xr.Dataset(
-            data_vars=arrays,
-            coords=coords
-        )
-        return self.s_topography
-
-    def set_values_to_sections(self,
-                               values: list,
-                               l0, l1,
-                               coords_base):
-        coords = coords_base
-        sections = self.grid.sections
-
-        for e, axis_coord in enumerate(sections.generate_axis_coord()):
-            resolution = sections.resolution[e]
-            l0_s = sections.length[e]
-            l1_s = sections.length[e + 1]
-            name, xy = axis_coord
-            coords['X'] = xy[:, 0]
-            coords['Y'] = xy[:, 1]
-
-            arrays = self.create_struc_xarrays(values, l0 + l0_s, l0 + l1_s,
-                                               resolution)
-
-            self.s_sections[name] = xr.Dataset(
-                data_vars=arrays,
-                coords=coords
-            )
-        return self.s_sections
-
-
-@_setdoc_pro(
-    [Grid.__doc__, Surfaces.__doc__, Series.__doc__, ds.weights_vector, ds.sfai,
-     ds.bai, ds.mai, ds.vai,
-     ds.lith_block, ds.sfm, ds.bm, ds.mm, ds.vm, ds.vertices, ds.edges,
-     ds.geological_map])
-class Solution(XSolution):
+class Solution(inheritance):
     """This class stores the output of the interpolation and the necessary objects
     to visualize and manipulate this data.
 
@@ -565,17 +273,11 @@ class Solution(XSolution):
         """
         self.vertices = []
         self.edges = []
-        if 'mask_topography' in kwargs:
-            mask_topography = kwargs.pop('mask_topography')
-        else:
-            mask_topography = True
 
-        if 'masked_marching_cubes' in kwargs:
-            masked_marching_cubes = kwargs.pop('masked_marching_cubes')
-        else:
-            masked_marching_cubes = True
+        mask_topography = kwargs.pop('mask_topography', True)
+        masked_marching_cubes = kwargs.pop('masked_marching_cubes', True)
 
-        self.padding_mask_matrix(mask_topography=mask_topography)
+        self. mask_matrix_pad = self.padding_mask_matrix(mask_topography=mask_topography)
         series_type = self.series.df['BottomRelation']
         s_n = 0
         active_indices = self.surfaces.df.groupby('isActive').groups[True]
@@ -583,11 +285,11 @@ class Solution(XSolution):
 
         # We loop the scalar fields
         for e, scalar_field in enumerate(self.scalar_field_matrix):
-            sfas = self.scalar_field_at_surface_points[e]
+
             # Drop
-            sfas = sfas[np.nonzero(sfas)]
-            mask_array = self.mask_matrix_pad[
-                e - 1 if series_type[e - 1] == 'Onlap' else e]
+            mask_array, sfas = self.prepare_marching_cubes_args(e,
+                                                                masked_marching_cubes,
+                                                                series_type)
             for level in sfas:
                 try:
                     v, s, norm, val = self.compute_marching_cubes_regular_grid(
@@ -598,12 +300,25 @@ class Solution(XSolution):
                                   '; Surface Number:' + str(s_n))
                     v = np.nan
                     s = np.nan
-
-                self.vertices.append(v)
-                self.edges.append(s)
-                idx = active_indices[s_n]
-                self.surfaces.df.loc[idx, 'vertices'] = [v]
-                self.surfaces.df.loc[idx, 'edges'] = [s]
-                s_n += 1
-
+                s_n = self.set_vertices_edges(active_indices, s, s_n, v)
         return self.vertices, self.edges
+
+    def set_vertices_edges(self, active_indices, s, s_n, v):
+        self.vertices.append(v)
+        self.edges.append(s)
+        idx = active_indices[s_n]
+        self.surfaces.df.loc[idx, 'vertices'] = [v]
+        self.surfaces.df.loc[idx, 'edges'] = [s]
+        s_n += 1
+        return s_n
+
+    def prepare_marching_cubes_args(self, e, masked_marching_cubes, series_type):
+
+        sfas = self.scalar_field_at_surface_points[e]
+        sfas = sfas[np.nonzero(sfas)]
+        if masked_marching_cubes is True:
+            mask_array = self.mask_matrix_pad[
+                e - 1 if series_type[e - 1] == 'Onlap' else e]
+        else:
+            mask_array = None
+        return mask_array, sfas
