+++ optional-requirements.txt
@@ -14,3 +14,5 @@ mplstereonet
 welly
 # gdal # install gdal with conda! `conda install gdal`
 
+# Subsurface
+subsurface
